+++ internal/clock/clock.go
@@ -1,0 +1,7 @@
+package clock
+
+import "time"
+
+type Clock interface {
+	Now() time.Time
+}
