+++ internal/clock/system_clock.go
@@ -1,0 +1,9 @@
+package clock
+
+import "time"
+
+type SystemClock struct{}
+
+func (SystemClock) Now() time.Time {
+	return time.Now()
+}
